@@ -3,20 +3,24 @@
 	"version": "0.1.0",
 	"private": true,
 	"dependencies": {
-		"@material-ui/core": "^4.11.0",
-		"@material-ui/icons": "^4.9.1",
+		"@babel/runtime": "7.0.0-beta.55",
+		"@emotion/react": "^11.14.0",
+		"@emotion/styled": "^11.14.0",
+		"@mui/icons-material": "^7.0.1",
+		"@mui/material": "^7.0.1",
 		"@testing-library/jest-dom": "^4.2.4",
 		"@testing-library/react": "^9.5.0",
 		"@testing-library/user-event": "^7.2.1",
-		"axios": "^0.21.1",
-		"firebase": "^8.2.3",
+		"@tweenjs/tween.js": "^25.0.0",
+		"axios": "^1.8.4",
+		"firebase": "^11.6.0",
 		"jquery": "^3.5.1",
-		"react": "^16.13.1",
-		"react-dom": "^16.13.1",
-		"react-flip-move": "^3.0.4",
-		"react-router-dom": "^5.2.0",
-		"react-scripts": "3.4.3",
-		"react-typing-animation": "^1.6.2"
+		"react": "^19.1.0",
+		"react-dom": "^19.1.0",
+		"react-router-dom": "^7.4.1",
+		"react-scripts": "5.0.1",
+		"react-simple-typewriter": "^5.0.1",
+		"three": "^0.175.0"
 	},
 	"scripts": {
 		"start": "react-scripts start",
